fix rename component StickyArrowButton to FloatingArrowButton - MEED-9761 - Meeds-io/MIPs#198
