+++ webapp/src/main/webapp/vue-apps/common/components/FloatingArrowButton.vue
